src/quant/monte_carlo.py
"""
Monte Carlo Simulation Engine for Probabilistic Risk Forecasts
Advanced quant-level simulation engine with multiple simulation strategies
"""
import asyncio
import logging
from typing import Dict, List, Optional, Tuple, Any, Union, Callable
from dataclasses import dataclass, field
from datetime import datetime, timedelta
from enum import Enum
import numpy as np
import pandas as pd
from scipy.stats import norm, gaussian_kde, percentileofscore
from scipy.optimize import minimize
import redis.asyncio as redis

from ..utils.performance_cache import PerformanceCache

logger = logging.getLogger(__name__)

class SimulationStrategy(Enum):
    """Available simulation strategies"""
    TRIANGULAR = "triangular"
    STATISTICAL = "statistical"
    EVENT_DRIVEN = "event_driven"
    CORRELATION = "correlation"
    MOMENTUM = "momentum"
    COMPOSITE = "composite"

class SamplingMethod(Enum):
    """Random sampling methods"""
    SIMPLE_RANDOM = "simple_random"
    LATIN_HYPERCUBE = "latin_hypercube"
    SOBOL = "sobol"
    IMPORTANCE_SAMPLING = "importance_sampling"

class SimulationMode(Enum):
    """Simulation execution modes"""
    SINGLE_PATH = "single_path"                    # Simulate one execution path
    MULTI_PATH = "multi_path"                      # Simulate all strategies
    PARALLEL_EXECUTION = "parallel_execution"      # Parallel path simulation
    SEQUENTIAL = "sequential"                       # Sequential path simulation

@dataclass
class MonteCarloResult:
    """Result of Monte Carlo simulation"""
    simulation_id: str
    strategy: SimulationStrategy
    num_simulations: int
    
    # Portfolio metrics (averages)
    expected_return: float
    return_distribution: np.ndarray  # Raw return samples
    sorted_returns: np.ndarray     # Sorted returns
    return_percentiles: np.ndarray  #  Return percentiles
    annualized_sharpe: float
    
    # Risk metrics
    var_95: float              # Value at Risk 95%
    var_99: float              # Value at Risk 99%
    cvar_95: float             # Conditional VaR 95%
    cvar_99: float             # Conditional VaR 99%
    probability_of_ruin: float     # Tail probability
    probability_of_profit: float   # Probability of profitable trade
    tail_expectation: float        # Expected tail expectation
    
    # Execution metrics
                avg_execution_time: float  # Average time per execution
        successful_executions: int
        failed_executions: int
        avg_slippage: float
        market_impact: float
        transaction_costs: float
    
    # Performance predictions
        best_scenario_return: float
        worst_scenario_return: float
        probability_exceeding_target: float
        confidence_interval: Dict[str, Tuple[float, float]]  # confidence intervals
    
    # Detailed distributions
    return_samples: List[Dict[str, Any]]
    execution_histories: List[Dict[str, Any]]
    time_series: List[Dict[str, Any]]
    
    timestamp: datetime = field(default_factory=datetime.now)
    optimization_time: float = 0.0
    metadata: Dict[str, Any] = field(default_factory=dict)
    total_trades: int = 0
    successful_trades: int = 0
    successful_trades: int = 0
    failed_trades: int = 0
    total_trades = 0
    successful_trades = 0
    failed_trades = 0
    successful_trades = 0
    successful_trades = 0
    successful_trades = 0
    successful_trades = successful_trades = 0
    successful_trades = 0
    successful_trades = successful_trades = 0
    successful_trades = 0
    successful_trades = successful_trades = 0
    successful_trades = successful_trades = 0
    successful_trades = successful_trades = 0
    successful_trades = 0
    successful_trades = 0
    total_trades = 0
    successful_trades = 0
    successful_trades = 0
    successful_trades = 0
    total_trades = 0
    successful_trades = 0
}