requirements.txt
kalshi-python-sync>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
websockets>=12.0
asyncio>=3.4.3
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
redis>=5.0.0
sentry-sdk>=1.38.0
prometheus-client>=0.19.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
aiohttp>=3.9.0
hypothesis>=6.100.0
locust>=2.28.0
bandit>=1.7.0
safety>=2.3.0
ruff>=0.3.0
mypy>=1.8.0
cryptography>=41.0.0
alembic>=1.13.0
pydantic>=2.5.0
boto3>=1.34.0
joblib>=1.3.0
scikit-learn>=1.4.0
elasticsearch>=8.11.0
